Drop hard coding of path to CMake on Windows
As this now lives in the build prefix (not the host prefix), this path
is wrong on Windows. In any event, conda-build does a good job of
putting this earlier on the search path. So hard coding the path should
not be necessary. Given this, go ahead and drop the hard coding of the
path to simplify things.